Formatting changes and changed type of recieved time-step to the hydro functions to float from double

+++ src/hydro/GizmoMFV/hydro.h
@@ -481,7 +481,7 @@ __attribute__((always_inline)) INLINE static void hydro_end_gradient(
  */
 __attribute__((always_inline)) INLINE static void hydro_prepare_force(
     struct part* restrict p, struct xpart* restrict xp,
-    const struct cosmology* cosmo, const double dt_alpha) {
+    const struct cosmology* cosmo, const float dt_alpha) {
 
   /* Initialise values that are used in the force loop */
   p->gravity.mflux[0] = 0.0f;
